+++ src/main/java/com/rviewer/skeletons/domain/dtos/response/DispenserResponse.java
@@ -1,5 +1,6 @@
 package com.rviewer.skeletons.domain.dtos.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.rviewer.skeletons.domain.dtos.request.DispenserRequest;
@@ -10,7 +11,7 @@ import lombok.Getter;
 public class DispenserResponse extends DispenserRequest {
   private UUID id;
 
-  public DispenserResponse(UUID id, float flowVolume) {
+  public DispenserResponse(UUID id, BigDecimal flowVolume) {
     super(flowVolume);
     this.id = id;
   }
